Remove dhcpcd
With Balena we ended up with 2 DHCP clients (NetworkManager and dhcpcd).

This resulted in one IP being obtained by NetworkManager which was visible
through ifconfig and another obtained by dhcpcd which didn't appear visible
through ifconfig.

In our reboot tests we noticed it exhausting a DHCP pool, this commit removes
dhcpcd to avoid that.